==> 12_sce_infected/14_venn_DE.Rmd <==
---
title: "Untitled"
author: "Kevin Rue-Albrecht"
date: "07/07/2018"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(cache = TRUE)
stopifnot(
  require(scater),
  require(scran),
  require(dplyr),
  require(ggrepel),
  require(RColorBrewer),
  require(venn)
)
sce.norm <- readRDS("rds/sce.norm.rds")
col.time <- brewer.pal(9, "Set3")[c(2,7:8)]
col.infection <- brewer.pal(12, "Paired")[c(9,4,2)]
col.status <- brewer.pal(12, "Paired")[c(9,1,7)]
names(col.time) <- levels(sce.norm$Time)[1:3]
names(col.infection) <- levels(sce.norm$Infection)
names(col.status) <- levels(sce.norm$Status)
outdir <- "14_Venn_DE_20180707_out"; dir.create(outdir, showWarnings = FALSE)
scde.group.res <- readRDS("rds/scde_group.rds")
scde.cluster.res <- readRDS("rds/scde_cluster.rds")
```

```{r}
convert.z.score <- function(x, one.sided = NULL) {
  z <- x$Z
  if(is.null(one.sided)) {
    pval = pnorm(-abs(z));
    pval = 2 * pval
  } else if(one.sided=="-") {
    pval = pnorm(z);
  } else {
    pval = pnorm(-z);
  }
    x <- cbind(
      x,
      p.value = pval
  )
  return(x);
}
```

# 4h

Venn of:

- cluster 2 vs 1 DE genes
- DE genes between D23580-infected and exposed 
- DE genes between LT2-infected and exposed

All three

```{r}
venn_sets <- list()
venn_sets[["cluster 2 vs 1"]] <- rownames(subset(
  convert.z.score(scde.cluster.res$`4h_cluster2-cluster1`),
  p.value < 0.01))
venn_sets[["STM-D23580"]] <- rownames(subset(
  convert.z.score(scde.group.res$`4h_STM-D23580_Infected-4h_STM-D23580_Exposed`),
  p.value < 0.01))
venn_sets[["STM-LT2"]] <- rownames(subset(
  convert.z.score(scde.group.res$`4h_STM-LT2_Infected-4h_STM-LT2_Exposed`),
  p.value < 0.01))
venn(venn_sets, zcolor = "style", cexil = 1, cexsn = 0.85)
```

D23580

```{r}
venn_sets <- list()
venn_sets[["cluster 2 vs 1"]] <- rownames(subset(
  convert.z.score(scde.cluster.res$`4h_cluster2-cluster1`),
  p.value < 0.01))
venn_sets[["STM-D23580"]] <- rownames(subset(
  convert.z.score(scde.group.res$`4h_STM-D23580_Infected-4h_STM-D23580_Exposed`),
  p.value < 0.01))
pdf(file.path(outdir, "4h_cluster2v1_D23580.pdf"), width = 5, height = 4)
venn(venn_sets, zcolor = c("darkorange3", col.infection["STM-D23580"]), cexil = 1, cexsn = 0.85)
dev.off()
```

Fisher's test

```{r}
bg <- union(
  rownames(scde.cluster.res$`4h_cluster2-cluster1`),
  rownames(scde.group.res$`4h_STM-D23580_Infected-4h_STM-D23580_Exposed`)
)
tmpTable <- data.frame(
    row.names = bg,
    cluster = bg %in% venn_sets[["cluster 2 vs 1"]],
    status = bg %in% venn_sets[["STM-D23580"]]
)
fisherTable <- with(tmpTable, table(cluster, status))
sink(file.path(outdir, "4h_cluster2v1_D23580_fisher.txt"))
fisher.test(fisherTable)
sink()
```

LT2

```{r}
venn_sets <- list()
venn_sets[["cluster 2 vs 1"]] <- rownames(subset(
  convert.z.score(scde.cluster.res$`4h_cluster2-cluster1`),
  p.value < 0.01))
venn_sets[["STM-LT2"]] <- rownames(subset(
  convert.z.score(scde.group.res$`4h_STM-LT2_Infected-4h_STM-LT2_Exposed`),
  p.value < 0.01))
pdf(file.path(outdir, "4h_cluster2v1_LT2.pdf"), width = 5, height = 4)
venn(venn_sets, zcolor = c("darkorange3", col.infection["STM-LT2"]), cexil = 1, cexsn = 0.85)
dev.off()
```

Fisher's test

```{r}
bg <- union(
  rownames(scde.cluster.res$`4h_cluster2-cluster1`),
  rownames(scde.group.res$`4h_STM-LT2_Infected-4h_STM-LT2_Exposed`)
)
tmpTable <- data.frame(
    row.names = bg,
    cluster = bg %in% venn_sets[["cluster 2 vs 1"]],
    status = bg %in% venn_sets[["STM-LT2"]]
)
fisherTable <- with(tmpTable, table(cluster, status))
sink(file.path(outdir, "4h_cluster2v1_LT2_fisher.txt"))
fisher.test(fisherTable)
sink()
```

# 6h

Venn of:

- cluster 2 vs 1 DE genes
- DE genes between D23580-infected and exposed 
- DE genes between LT2-infected and exposed

All three

```{r}
venn_sets <- list()
venn_sets[["cluster 2 vs 1"]] <- rownames(subset(
  convert.z.score(scde.cluster.res$`6h_cluster2-cluster1`),
  p.value < 0.01))
venn_sets[["STM-D23580"]] <- rownames(subset(
  convert.z.score(scde.group.res$`6h_STM-D23580_Infected-6h_STM-D23580_Exposed`),
  p.value < 0.01))
venn_sets[["STM-LT2"]] <- rownames(subset(
  convert.z.score(scde.group.res$`6h_STM-LT2_Infected-6h_STM-LT2_Exposed`),
  p.value < 0.01))
venn(venn_sets, zcolor = "style", cexil = 1, cexsn = 0.85)
```

D23580

```{r}
venn_sets <- list()
venn_sets[["cluster 2 vs 1"]] <- rownames(subset(
  convert.z.score(scde.cluster.res$`6h_cluster2-cluster1`),
  p.value < 0.01))
venn_sets[["STM-D23580"]] <- rownames(subset(
  convert.z.score(scde.group.res$`6h_STM-D23580_Infected-6h_STM-D23580_Exposed`),
  p.value < 0.01))
pdf(file.path(outdir, "6h_cluster2v1_D23580.pdf"), width = 5, height = 4)
venn(venn_sets, zcolor = c("darkorange3", col.infection["STM-D23580"]), cexil = 1, cexsn = 0.85)
dev.off()
```

Fisher's test

```{r}
bg <- union(
  rownames(scde.cluster.res$`6h_cluster2-cluster1`),
  rownames(scde.group.res$`6h_STM-D23580_Infected-6h_STM-D23580_Exposed`)
)
tmpTable <- data.frame(
    row.names = bg,
    cluster = bg %in% venn_sets[["cluster 2 vs 1"]],
    status = bg %in% venn_sets[["STM-D23580"]]
)
fisherTable <- with(tmpTable, table(cluster, status))
sink(file.path(outdir, "6h_cluster2v1_D23580_fisher.txt"))
fisher.test(fisherTable)
sink()
```

LT2

```{r}
venn_sets <- list()
venn_sets[["cluster 2 vs 1"]] <- rownames(subset(
  convert.z.score(scde.cluster.res$`6h_cluster2-cluster1`),
  p.value < 0.01))
venn_sets[["STM-LT2"]] <- rownames(subset(
  convert.z.score(scde.group.res$`6h_STM-LT2_Infected-6h_STM-LT2_Exposed`),
  p.value < 0.01))
pdf(file.path(outdir, "6h_cluster2v1_LT2.pdf"), width = 5, height = 4)
venn(venn_sets, zcolor = c("darkorange3", col.infection["STM-LT2"]), cexil = 1, cexsn = 0.85)
dev.off()
```

Fisher's test

```{r}
bg <- union(
  rownames(scde.cluster.res$`6h_cluster2-cluster1`),
  rownames(scde.group.res$`6h_STM-LT2_Infected-6h_STM-LT2_Exposed`)
)
tmpTable <- data.frame(
    row.names = bg,
    cluster = bg %in% venn_sets[["cluster 2 vs 1"]],
    status = bg %in% venn_sets[["STM-LT2"]]
)
fisherTable <- with(tmpTable, table(cluster, status))
sink(file.path(outdir, "6h_cluster2v1_LT2_fisher.txt"))
fisher.test(fisherTable)
sink()
```
